requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.114.2
filelock==3.16.0
fsspec==2024.9.0
h11==0.14.0
idna==3.10
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
openai-whisper==20231117
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
pydub==0.25.1
python-multipart==0.0.9
regex==2024.9.11
requests==2.32.3
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
starlette==0.38.5
sympy==1.13.2
tiktoken==0.7.0
torch==2.4.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
